New quality metric
The quality metric uses the following color scheme:

-7 dB dropout point (red!)
-6 is meh data point (yellow!)
-5 is OK data point (yellow green!)
-4.3 dB 100% good data point (green!)

The metric is a running average of the (abs) phase angle after the PLL but before the filter. It is basically measuring how 'tight' the constellation is, which I have empirically tied to the demodulation quality. 

The idea of this realtime (and color coded!) metric is to help the end user develop a better system, be it antenna pointing, antenna gain, noise floor (IE LNA) or optimizing dynamic gain settings on the receiver.